feat: index.jsをsrcからルートへ移動させた
