+++ localRun.js
@@ -1,0 +1,5 @@
+const index = require('./index.js');
+
+(async () => {
+  await index.handler();
+})();
